Web Interface for DNS external resolvers application.
