+++ applications/detection/dns_external_resolvers/spark/modules/converter.py
