+++ applications/detection/dns_external_resolvers/web-interface/controllers/dns_external_resolvers.py
@@ -11,14 +11,14 @@ from global_functions import escape
 #----------------- Main Functions -------------------#
 
 
-def dns_statistics():
+def dns_external_resolvers():
     """
     Show the main page of the DNS statistics section.
 
     :return: Empty dictionary
     """
     # Use standard view
-    response.view = request.controller + '/dns_statistics.html'
+    response.view = request.controller + '/dns_external_resolvers.html'
     return dict()
 
 
@@ -38,7 +38,7 @@ def get_top_n_statistics():
         return json_response
 
     # Check mandatory inputs
-    if not (request.get_vars.beginning and request.get_vars.end and request.get_vars.type and request.get_vars.number):
+    if not (request.get_vars.beginning and request.get_vars.end and request.get_vars.type and request.get_vars.number and request.get_vars.filter):
         json_response = '{"status": "Error", "data": "Some mandatory argument is missing!"}'
         return json_response
 
@@ -46,54 +46,44 @@ def get_top_n_statistics():
     beginning = escape(request.get_vars.beginning)
     end = escape(request.get_vars.end)
     type = escape(request.get_vars.type)
+    filter = escape(request.get_vars.filter)
     number = int(escape(request.get_vars.number))
 
     try:
-        # Elastic query
         client = elasticsearch.Elasticsearch([{'host': myconf.get('consumer.hostname'), 'port': myconf.get('consumer.port')}])
         elastic_bool = []
         elastic_bool.append({'range': {'@timestamp': {'gte': beginning, 'lte': end}}})
-        elastic_bool.append({'term': {'@stat_type': type}})
+        elastic_bool.append({'term': {'@type': 'external_dns_resolver'}})
 
+        # Set filter
+        if filter != 'none':
+            elastic_should = []
+            elastic_should.append({'term': {'src_ip.raw': filter}})
+            elastic_should.append({'term': {'resolver_ip.raw': filter}})
+            elastic_bool.append({'bool': {'should': elastic_should}})
         # Prepare query
         qx = Q({'bool': {'must': elastic_bool}})
 
-        # Set query according to the statistic type
-        if type == "queried_by_ip":
-            search_ip = Search(using=client, index='_all').query(qx)
-            search_ip.aggs.bucket('all_nested', 'nested', path='data_array') \
-                .bucket('by_key', 'terms', field='data_array.key.raw', size=2147483647)\
-                .bucket('by_ip', 'terms', field='data_array.ip.raw', size=1, order={'sum_by_ip': 'desc'}) \
-                .bucket('sum_by_ip', 'sum', field='data_array.value')
-            search_ip.aggs['all_nested']['by_key'].bucket('sum_total', 'sum', field='data_array.value')
-            results = search_ip.execute()
-        else:
-            search_ip = Search(using=client, index='_all').query(qx)
-            search_ip.aggs.bucket('all_nested', 'nested', path='data_array') \
-                .bucket('by_key', 'terms', field='data_array.key.raw', size=2147483647) \
-                .bucket('stats_sum', 'sum', field='data_array.value')
-            results = search_ip.execute()
+        # Get ordered data (with maximum size aggregation)
+        search = Search(using=client, index='_all').query(qx)
+        search.aggs.bucket('by_src', 'terms', field='src_ip.raw', size=2147483647)\
+              .bucket('by_dst', 'terms', field='resolver_ip.raw', size=2147483647)\
+              .bucket('top_src_dst', 'top_hits', size=1, sort=[{'@timestamp': {'order': 'desc'}}])
+        results = search.execute()
 
-        # Prepare data variable
-        data = ""
         # Prepare ordered collection
         counter = collections.Counter()
+        for src_buckets in results.aggregations.by_src.buckets:
+            if type == "sources":
+                counter[src_buckets.key] = len(src_buckets.by_dst.buckets)
+            else:
+                for dst_buckets in src_buckets.by_dst.buckets:
+                    counter[dst_buckets.key] += 1
 
-        if type == "queried_by_ip":
-            for record in results.aggregations.all_nested.by_key.buckets:
-                top_ip = record.by_ip.buckets[0]
-                counter[(record.key, top_ip.key, int(top_ip.sum_by_ip.value))] = int(record.sum_total.value)
-
-            # Select top N (number) values
-            for value, count in counter.most_common(number):
-                data += value[0] + "," + value[1] + "," + str(value[2]) + "," + str(count) + ","
-        else:
-            for all_buckets in results.aggregations.all_nested.by_key:
-                counter[all_buckets.key] += int(all_buckets.stats_sum.value)
-
-            # Select top N (number) values
-            for value, count in counter.most_common(number):
-                data += value + "," + str(count) + ","
+        # Select first N (number) values
+        data = ""
+        for ip, count in counter.most_common(number):
+            data += ip + "," + str(count) + ","
 
         # Remove trailing comma
         data = data[:-1]
@@ -122,14 +112,14 @@ def get_records_list():
         return json_response
 
     # Check mandatory inputs
-    if not (request.get_vars.beginning and request.get_vars.end and request.get_vars.type):
+    if not (request.get_vars.beginning and request.get_vars.end and request.get_vars.filter):
         json_response = '{"status": "Error", "data": "Some mandatory argument is missing!"}'
         return json_response
 
     # Parse inputs and set correct format
     beginning = escape(request.get_vars.beginning)
     end = escape(request.get_vars.end)
-    type = escape(request.get_vars.type)
+    filter = escape(request.get_vars.filter)
 
     try:
         # Elastic query
@@ -137,23 +127,30 @@ def get_records_list():
             [{'host': myconf.get('consumer.hostname'), 'port': myconf.get('consumer.port')}])
         elastic_bool = []
         elastic_bool.append({'range': {'@timestamp': {'gte': beginning, 'lte': end}}})
-        elastic_bool.append({'term': {'@stat_type': type}})
+        elastic_bool.append({'term': {'@type': 'external_dns_resolver'}})
 
-        # Prepare query
+        # Set filter
+        if filter != 'none':
+            elastic_should = []
+            elastic_should.append({'term': {'src_ip.raw': filter}})
+            elastic_should.append({'term': {'resolver_ip.raw': filter}})
+            elastic_bool.append({'bool': {'should': elastic_should}})
         qx = Q({'bool': {'must': elastic_bool}})
 
-        # Set query according to the statistic type
-        search_ip = Search(using=client, index='_all').query(qx)
-        search_ip.aggs.bucket('all_nested', 'nested', path='data_array')\
-            .bucket('by_key', 'terms', field='data_array.key.raw', size=2147483647)\
-            .bucket('stats_sum', 'sum', field='data_array.value')
-        results = search_ip.execute()
+        # Search with maximum size aggregations
+        search = Search(using=client, index='_all').query(qx)
+        search.aggs.bucket('by_src', 'terms', field='src_ip.raw', size=2147483647)\
+              .bucket('by_dst', 'terms', field='resolver_ip.raw', size=2147483647)\
+              .bucket('top_src_dst', 'top_hits', size=1, sort=[{'@timestamp': {'order': 'desc'}}])
+        results = search.execute()
 
+        # Result Parsing into CSV in format: timestamp, source_ip, destination_ip, flows, duration
         data = ""
-        for all_buckets in results.aggregations.all_nested.by_key:
-            data += all_buckets.key + "," + str(int(all_buckets.stats_sum.value)) + ","
-
-        # Remove trailing comma
+        for src_aggregations in results.aggregations.by_src.buckets:
+            for result in src_aggregations.by_dst.buckets:
+                record = result.top_src_dst.hits.hits[0]["_source"]
+                data += record["timestamp"].replace("T", " ").replace("Z", "") + "," + record["src_ip"] + "," \
+                    + record["resolver_ip"] + "," + str(record["flows_total"]) + ","
         data = data[:-1]
 
         json_response = '{"status": "Ok", "data": "' + data + '"}'
