+++ applications/detection/dns_external_resolvers/web-interface/views/menu/dns_external_resolvers.html
@@ -1,0 +1,4 @@
+<!-- Protocols Statistics -->
+<li>
+    <a href="/dns-external-resolvers/dns-external-resolvers" id="menu-dns-external-resolvers">DNS External Resolvers</a>
+</li>
