--- applications/statistics/dns_statistics/web-interface/views/menu/dns_statistics.html
@@ -1,4 +1,0 @@
-<!-- Protocols Statistics -->
-<li>
-    <a href="/dns-statistics/dns-statistics" id="menu-dns-statistics">DNS Statistics</a>
-</li>
